Add a controller for serving files.
The FileController adds Content-type and Content-disposition
headers to the response, as retrieved from the filestore.

It is meant for use in development. Presumably, in production, the http
server will be able to handle the job.

Note that, with this change, the files/ directory cannot be in
floof/public/, or pylons' static file route will take precedence.