Add controls for changing display name.
Start a "User Info" tab to put them in.

Currently, display name validation rejects anything that isn't printable
ASCII, and reduction (to determine whether a display name is trivially
reducible to the user's username) consists only of converting to
lowercase and replacing runs of non-word-characters with a single
underscore.  Both methods should be easy enough to expand, though.