Use display name everywhere.
user_link() accounts for most appearances, and displays the username
alongside the display name if the display name isn't trivial.  User
pages display username as the main name at the top, with display name
shown second, if any; both names are shown even if the display name is
trivial.  The rest of the page then uses just the display name.