Remove the last of old Pylons configuration.  #441
This adds a new file paster.ini which should be appropriate for running
the app either in dev or prod.  Hopefully.  Eh.

Also includes a fix for static files and beaker configuration, since
those were both in the old deployment template.

And upgrades us to Pyramid 1.1.  Which is completely unrelated.