Remove floof-specific logging.  :(
This was a bad idea and I apologize.  It's made even badder by the
transaction stuff baked into Pyramid, which makes extra-transactional
logging even harder to implement, *and* I'm not convinced the kinds of
things in the admin log should even go there outside a transaction.

Regular app logging will go through the logging module as normal.  Once
there are admin things that are worth logging, we can add some sugar to
do that, but it'll be a separate mechanism.