Make stashing suck marginally less.
Most notably, the automatic stash retrieval will now turn a GET request
into a POST if the stash used a key (rather than the immediate flag) and
there is data in the stash's `post` attribute.  This removes the ability to
review an action after re-auth, but also removes the need work out how to
GET a POST resource, which may have different requirements for each view,
and then communicate the solution to manage_stashes().