Fix bug on verification of stashed OpenID requests.
When the stash is followed, it is appended to the URL.  This is necessary
for POST stashes so that the POST data can be safely retrieved but it is
also necessary for POST and "GET" stashes alike so that the stash may be
deleted and stop cluttering the session.

For OpenID this is a problem as it verifies the full URL against the
(signed) openid.return_to parameter, so we need to strip it before passing
the current params to the OpenID verifier.

Note that return_key params will only be stripped if none of them are
valid (ie are not backed by actual stashes), otherwise they are kept and
the verification process is allowed to fail.